Builds with an angry error check from clang and no static analysis errors. Added Factories and added a damage spell factory as an example. Added tests for factories. Added death rattle and on going effects lists.

+++ deck.cpp
@@ -92,9 +92,9 @@ void deck::insert(const std::string& filename)
     }
 }
 
-int deck::count()
+unsigned int deck::count()
 {
-    return m_cards.size();
+    return static_cast<unsigned int>(m_cards.size());
 }
 
 bool deck::can_draw()
